backend/pyproject.toml
[project]
name = "hackfest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "argon2-cffi>=25.1.0",
    "dotenv>=0.9.9",
    "fastapi[all]>=0.127.0",
    "google>=3.0.0",
    "google-generativeai>=0.8.6",
    "jose>=1.0.0",
    "numpy>=2.2.6",
    "openmeteo-requests>=1.7.4",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "sqlalchemy>=2.0.45",
    "tensorflow>=2.20.0",
]
